requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asyncio==4.0.0
attrs==25.3.0
audioread==3.0.1
backoff==2.2.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
fonttools==4.59.2
frozenlist==1.7.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
jiter==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-openai==0.3.32
langchain-text-splitters==0.3.11
langfuse==3.3.4
langsmith==0.4.23
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
marshmallow==3.26.1
matplotlib==3.10.6
moviepy==2.2.1
msgpack==1.1.1
multidict==6.6.4
mypy_extensions==1.1.0
noisereduce==3.0.3
numba==0.61.2
numpy==2.2.6
openai==1.104.2
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.3
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
proglog==0.1.12
propcache==0.3.2
protobuf==6.32.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
SQLAlchemy==2.0.43
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
zstandard==0.24.0
